refactor: Take the CardComponent out of the mutation components
