+++ apps/dashboard/src/components/mutations/MutationsUserTabs.vue
@@ -1,16 +1,24 @@
 <template>
         <TabView class="w-full">
             <TabPanel header="PH-Balance mutations">
-                <MutationsBalanceCard
-                    :paginator="paginator"
-                    :getMutations="props.getBalanceMutations"
-                />
+                <CardComponent
+                    header="PH-Recent transactions"
+                >
+                    <MutationsBalance
+                        :paginator="paginator"
+                        :getMutations="props.getBalanceMutations"
+                    />
+                </CardComponent>
             </TabPanel>
             <TabPanel header="PH-Created transactions">
-                <MutationsSellerCard
-                    :paginator="paginator"
-                    :getMutations="props.getSellerMutations"
-                />
+                <CardComponent
+                     header="PH-Created transactions"
+                >
+                    <MutationsSeller
+                        :paginator="paginator"
+                        :getMutations="props.getSellerMutations"
+                    />
+                </CardComponent>
             </TabPanel>
         </TabView>
 </template>
@@ -18,8 +26,9 @@
 import TabView from 'primevue/tabview';
 import TabPanel from 'primevue/tabpanel';
 
-import MutationsBalanceCard from '@/components/mutations/MutationsBalanceCard.vue';
-import MutationsSellerCard from '@/components/mutations/MutationsSellerCard.vue';
+import MutationsBalance from '@/components/mutations/MutationsBalance.vue';
+import MutationsSeller from '@/components/mutations/MutationsSeller.vue';
+import CardComponent from "@/components/CardComponent.vue";
 
 const props = defineProps({
     paginator: {
